fix: save photo at filtering and refactoring
