+++ features/posts/PostDescriptionModal.tsx
@@ -1,7 +1,9 @@
 import { ImageToolModal } from "common/hoc/ImageToolModal"
 import { useState } from "react"
 import { styled } from "styled-components"
-import { PhotoType } from "./PostPhotoEditorModal"
+import { PhotoType } from "./PostCreationModal"
+import { useCreatePostMutation } from "assets/store/api/posts/postsApi"
+import Canvas from "./Canvas"
 
 const PostDescriptionModal = ({
         handleBackToFilters, 
@@ -14,18 +16,66 @@ const PostDescriptionModal = ({
 }) => {
 
     const [photo, setPhoto] = useState(photoPost[0])
+    const [canvasPhoto, setCanvasPhoto] = useState<string []>([])
+    const [createPostHandler] = useCreatePostMutation();
 
-    console.log(photoPost)
+   
 
     const handleBack = () => {
         handleBackToFilters(photoPost)
     }
 
-    const handlePublishButton = () => {
-        alert('send request')
+    const handleCanvas = (photoUrl: string) => {
+        const newList = [...canvasPhoto, photoUrl]
+        setCanvasPhoto(newList)
     }
 
-    return (
+    const handlePublishButton = () => {
+        console.log(canvasPhoto)
+
+  //   const formData = new FormData();
+  //   formData.append("description", "dsgasdg dsagsda gsda g");
+  //   photoPost.map((photo) => formData.append("posts", photo as File));
+
+  //   createPostHandler(formData);
+  // };
+
+  // Сохранение значений в локальный state при перемещении бегунка
+  // const handleSlider = (setState: (arg: number) => void) => (e: React.ChangeEvent<HTMLInputElement>) => {
+  //     if (e.target) {
+  //       setState(parseInt(e.target.value));
+  //     }
+  //   };
+
+  // // Обработчик сохранени отредактированного изображения
+  // const handleSave = async () => {
+  //   // подготовка данных
+  //   if (cropRef.current) {
+  //     const avatar = cropRef.current.getImage().toDataURL();
+
+  //     // преобразование base64 в file
+  //     const result = await fetch(avatar);
+  //     const blob = await result.blob();
+  //     const file = new File([blob], "avatar", {type: "image/png"});
+
+  //     // преобразование file в FormData
+  //     const formData = new FormData();
+  //     formData.append("avatar", file as File);
+  //     // setPhotoPost((prev) => [...prev, file]);
+
+  //     const newList = [...photoPost, {photoUrl: avatar, filter: ''}]
+  //     setPhotoPost(newList)
+  //     debugger
+  //     try {
+
+  //     } catch (error) {
+  //       console.log(error);
+  //     }
+  //   }
+  // };
+    }
+
+    return (<>
         <ImageToolModal 
             handleModalClose={handleModalClose} 
             photoPost={photoPost} 
@@ -40,6 +90,17 @@ const PostDescriptionModal = ({
                                 <div>Description</div>
                 </StyledDescriptionContainer>
         </ImageToolModal>
+        {photoPost.map( (el, index) => 
+            <Canvas 
+                key={index} 
+                photo={el.photoUrl}
+                filter={el.filter} 
+                width={"0px"}
+                height={"0px"} 
+                setImageUrl={handleCanvas}
+            />
+          )}
+          </>
     )
 }
